requirements.txt
streamlit 
huggingface_hub 
langchain
requests
python-dotenv